Refactor to put all data creation under the entities folder.

+++ plangs/main.py
@@ -1,30 +1,46 @@
-from pprint import pprint
-import kuzu
-import shutil
+import importlib
 import os
+import shutil
+from glob import iglob
 
-from langs.python import LangPython
+import kuzu
 
-ROOT_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..")
+from plangs.decorators import invoke_entity_creators, invoke_relationship_creators
+from plangs.entities import IdLang
+
+ROOT_DIR: str = os.path.abspath(
+    os.path.join(os.path.dirname(os.path.abspath(__file__)), "..")
+)
 
 
-def createDB() -> kuzu.Database:
-    """Create a new database. Removes the old one if it exists."""
+def createDB(remove_existing: bool = True) -> kuzu.Database:
+    """Create a new database."""
     path = os.path.join(ROOT_DIR, "db")
-    shutil.rmtree(path, ignore_errors=True)
+    if remove_existing:
+        shutil.rmtree(path, ignore_errors=True)
     db = kuzu.Database(path, buffer_pool_size=1024**3)
     return db
 
 
-def createConnection(db: kuzu.Database) -> kuzu.Connection:
-    return kuzu.Connection(db)
+def load_all_entities():
+    """
+    Import every module under the entities directory.
+    This should populate all the functions that define entities and relationships.
+    """
+    path = os.path.join(ROOT_DIR, "plangs/entities/**/*.py")
+    for modpath in iglob(path):
+        mod = modpath.split(ROOT_DIR)[-1].replace("/", ".").replace("\\", ".")[1:-3]
+        importlib.import_module(mod)
 
 
 if __name__ == "__main__":
-    conn = createConnection(createDB())
+    # conn = createConnection(createDB())
 
-    with open(os.path.join(ROOT_DIR, "cypher/schema.cypher"), "r") as schema:
-        conn.execute(schema.read())
+    # with open(os.path.join(ROOT_DIR, "cypher/schema.cypher"), "r") as schema:
+    #     conn.execute(schema.read())
 
-    pprint(LangPython.model_dump())
+    load_all_entities()
+    invoke_entity_creators()
+    invoke_relationship_creators()
 
+    print(IdLang.PYTHON.get().model_dump_json(indent=2))
